third_party/WebKit/LayoutTests/fast/canvas-api/canvas-getImageData-invalid.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<body>
<script>
test(function(t) {
    ctx = document.createElement('canvas').getContext('2d');

    assert_throws("INDEX_SIZE_ERR", function() {
      ctx.getImageData(10, 10, 0, 10);
    });
    assert_throws("INDEX_SIZE_ERR", function() {
      ctx.getImageData(10, 10, 10, 0);
    });
}, "Test the handling of invalid arguments in canvas getImageData().");
</script>
</body>
